estavamos a ler o recipe como float em vez de double :/, daí os erros de precisam ...

+++ trabalho-pratico/src/entities/artists.c
@@ -111,7 +111,7 @@ Artist create_artist_from_tokens(char **tokens){
     if (valid)
     { // negação do caso inválido (individual com id e constituent ñ nulo)
         int id = atoi(tokens[0] + 1);
-        float recipe_per_stream = strtod(tokens[3], NULL);
+        double recipe_per_stream = strtod(tokens[3], NULL);
         a = create_art(id, tokens[1], recipe_per_stream, id_constituent, tokens[5], art_type);
     }
     else if (id_constituent != NULL)
